# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API системы опросов для преддипломной практики",
        "contact": {
            "email": "vadim.starostin.07@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
        "/api/surveys/{id}/answers": {
            "post": {
                "tags": [
                    "Answers"
                ],
                "summary": "Пройти опрос",
                "operationId": "6a569a9a2ac7884ee4aeebd3a5d256b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "integer"
                                                },
                                                "option_id": {
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "text_answer": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Сохранено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/surveys/{id}/stats": {
            "get": {
                "tags": [
                    "Answers"
                ],
                "summary": "Статистика опроса (публичная)",
                "operationId": "5f53fc0c46baa25bfae67fd7c3caa07a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON со статистикой"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация нового пользователя",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Vadim"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешно"
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация (получение токена)",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Токен получен"
                    },
                    "401": {
                        "description": "Неверные данные"
                    }
                }
            }
        },
        "/api/surveys/{surveyId}/questions": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Добавить новый вопрос в опрос",
                "operationId": "6f9771f9c1db076c4fa171d414437ea3",
                "parameters": [
                    {
                        "name": "surveyId",
                        "in": "path",
                        "description": "ID опроса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "type"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Какой тип тюнинга вас интересует?"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "text",
                                            "radio",
                                            "checkbox"
                                        ]
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Вопрос успешно добавлен"
                    },
                    "403": {
                        "description": "Опрос уже опубликован или закрыт"
                    },
                    "404": {
                        "description": "Опрос не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/questions/{id}": {
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Обновить текст вопроса",
                "operationId": "795b02e5ecdc23fd74f20e0671a40f8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID вопроса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Обновленный текст вопроса"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Вопрос успешно обновлен"
                    },
                    "403": {
                        "description": "Доступ запрещен или опрос не в режиме черновика"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}